Refactor core blocks to have deprecated extracted to their ownf files (p.1)
