+++ packages/block-library/src/button/index.js
@@ -1,33 +1,20 @@
 /**
- * External dependencies
- */
-import { omit, pick } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { __, _x } from '@wordpress/i18n';
-import { RichText } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
+import deprecated from './deprecated';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
 import save from './save';
 
-const { name, attributes: blockAttributes } = metadata;
+const { name } = metadata;
 
 export { metadata, name };
-
-const colorsMigration = ( attributes ) => {
-	return omit( {
-		...attributes,
-		customTextColor: attributes.textColor && '#' === attributes.textColor[ 0 ] ? attributes.textColor : undefined,
-		customBackgroundColor: attributes.color && '#' === attributes.color[ 0 ] ? attributes.color : undefined,
-	}, [ 'color', 'textColor' ] );
-};
 
 export const settings = {
 	title: __( 'Button' ),
@@ -45,77 +32,5 @@ export const settings = {
 	],
 	edit,
 	save,
-	deprecated: [ {
-		attributes: {
-			...pick( blockAttributes, [ 'url', 'title', 'text' ] ),
-			color: {
-				type: 'string',
-			},
-			textColor: {
-				type: 'string',
-			},
-			align: {
-				type: 'string',
-				default: 'none',
-			},
-		},
-
-		save( { attributes } ) {
-			const { url, text, title, align, color, textColor } = attributes;
-
-			const buttonStyle = {
-				backgroundColor: color,
-				color: textColor,
-			};
-
-			const linkClass = 'wp-block-button__link';
-
-			return (
-				<div className={ `align${ align }` }>
-					<RichText.Content
-						tagName="a"
-						className={ linkClass }
-						href={ url }
-						title={ title }
-						style={ buttonStyle }
-						value={ text }
-					/>
-				</div>
-			);
-		},
-		migrate: colorsMigration,
-	},
-	{
-		attributes: {
-			...pick( blockAttributes, [ 'url', 'title', 'text' ] ),
-			color: {
-				type: 'string',
-			},
-			textColor: {
-				type: 'string',
-			},
-			align: {
-				type: 'string',
-				default: 'none',
-			},
-		},
-
-		save( { attributes } ) {
-			const { url, text, title, align, color, textColor } = attributes;
-
-			return (
-				<div className={ `align${ align }` } style={ { backgroundColor: color } }>
-					<RichText.Content
-						tagName="a"
-						href={ url }
-						title={ title }
-						style={ { color: textColor } }
-						value={ text }
-					/>
-				</div>
-			);
-		},
-		migrate: colorsMigration,
-	},
-	],
+	deprecated,
 };
